Add files via upload
So button should be fixed and the master and results pages are now linked by the button. Maybe not the best way of doing it but for the sake of having it ready for the sprint next week should be fine. create a default react project,  install react.js, install material.ui, and install react-router-dom@6. in public folder create images folder for the image. everything else goes in the src folder also in the src folder create a pages folder for the results.js file.